pyproject.toml
[project]
name = "socfusion"
version = "0.1.0"
description = "AI agent for automating tasks and decision-making in SOC"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.13",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "async-timeout==4.0.3",
    "attrs==25.1.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "click-default-group==1.2.4",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.9",
    "frozenlist==1.5.0",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.20",
    "langchain-core==0.3.43",
    "langchain-text-splitters==0.3.6",
    "langsmith==0.3.11",
    "multidict==6.1.0",
    "numpy==2.2.3",
    "orjson==3.10.15",
    "packaging==24.2",
    "pip==22.0.2",
    "playwright==1.50.0",
    "propcache==0.3.0",
    "psycopg2==2.9.10",
    "pyaes==1.6.1",
    "pyasn1==0.6.1",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pyee==12.1.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rsa==4.9",
    "setuptools==59.6.0",
    "shot-scraper==1.5",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.38",
    "starlette==0.45.3",
    "telethon==1.38.1",
    "tenacity==9.0.0",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "yarl==1.18.3",
    "zstandard==0.23.0",
    "cffi==1.17.1",
    "cryptography==44.0.2",
    "pycparser==2.22",
    "pyspnego==0.11.2",
    "smbprotocol==1.15.0",
    "pypdf2==3.0.1",
    "alembic==1.15.1",
    "mako==1.3.9",
    "grpcio==1.70.0",
    "grpcio-tools==1.70.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "hyperframe==6.1.0",
    "portalocker==2.10.1",
    "protobuf==5.29.3",
    "qdrant-client==1.13.3",
    "redis==5.2.1",
    "blinker==1.9.0",
    "community==1.0.0b1",
    "flask==3.1.0",
    "itsdangerous==2.2.0",
    "werkzeug==3.1.3",
    "dataclasses-json==0.6.7",
    "httpx-sse==0.4.0",
    "langchain-community==0.3.19",
    "marshmallow==3.26.1",
    "mypy-extensions==1.0.0",
    "typing-inspect==0.9.0",
    "dynaconf==3.2.10",
    "aiosqlite==0.21.0",
    "apprise==1.9.2",
    "asgi-lifespan==2.1.0",
    "asyncpg==0.30.0",
    "cachetools==5.5.2",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "coolname==2.2.0",
    "dateparser==1.2.1",
    "deprecated==1.2.18",
    "docker==7.1.0",
    "graphviz==0.20.3",
    "griffe==1.6.0",
    "humanize==4.12.1",
    "importlib-metadata==8.6.1",
    "jinja2-humanize-extension==0.4.0",
    "markdown==3.7",
    "oauthlib==3.2.2",
    "opentelemetry-api==1.31.0",
    "pathspec==0.12.1",
    "pendulum==3.0.0",
    "prefect==3.2.12",
    "prometheus-client==0.21.1",
    "pydantic-extra-types==2.10.3",
    "python-slugify==8.0.4",
    "python-socks==2.7.1",
    "readchar==4.2.1",
    "requests-oauthlib==2.0.0",
    "rfc3339-validator==0.1.4",
    "ruamel-yaml==0.18.10",
    "ruamel-yaml-clib==0.2.12",
    "text-unidecode==1.3",
    "time-machine==2.16.0",
    "toml==0.10.2",
    "tzlocal==5.3.1",
    "ujson==5.10.0",
    "uv==0.6.6",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "zipp==3.21.0",
    "filelock==3.18.0",
    "huggingface-hub==0.29.3",
    "safetensors==0.5.3",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "torch>=2.6.0",
    "sentence-transformers>=3.4.1",
    "docling>=2.15.1",
    "fastmcp>=0.4.1",
    "mistune>=3.1.3",
    "python-multipart>=0.0.20",
]

[tool.dynaconf]
settings_files = [".env"]  
load_dotenv = true  
environments = true  

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.10",
    "alembic>=1.15.1",
    "docling>=2.15.1",
    "dynaconf>=3.2.10",
    "fastapi>=0.115.9",
    "langchain-community>=0.3.19",
    "prefect>=3.2.12",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.13.3",
    "redis>=5.2.1",
    "sentence-transformers>=3.4.1",
    "smbprotocol>=1.15.0",
    "sqlalchemy>=2.0.38",
    "transformers>=4.49.0",
    "uv>=0.6.6",
    "uvicorn>=0.34.0",
]
fastapi = [
    "python-multipart>=0.0.20",
]
fastmcp = [
    "fastapi>=0.115.9",
    "fastmcp>=0.4.1",
    "uv>=0.6.6",
    "uvicorn>=0.34.0",
]
worker = [
    "aiohttp>=3.11.13",
    "prefect>=3.2.12",
    "python-multipart>=0.0.20",
]
